finished pattern mvp

+++ pattern/Point.java
@@ -66,7 +66,7 @@ public class Point implements Comparable<Point> {
             return Double.POSITIVE_INFINITY;
         if (this.y == that.y)
             return +0.0;
-        return (this.y - that. y) / (this.x - that.x);
+        return (this.y - (double)that. y) / (this.x - that.x);
 
     }
